requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dateutil>=2.8.2